Reorganize and cleaning code
