+++ src/app/app.component.html
@@ -1,8 +1,8 @@
-<div *ngIf="isLoggedIn(); else showLogIn">
+<ng-container *ngIf="isLoggedIn(); else showLogIn">
   <mat-card>
     <app-library></app-library>
   </mat-card>
-</div>
-<ng-template #showLogIn
-  ><app-login></app-login>
+</ng-container>
+<ng-template #showLogIn>
+  <app-login></app-login>
 </ng-template>
