+++ src/app/app.module.ts
@@ -1,63 +1,60 @@
 import { NgModule } from '@angular/core';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { BrowserModule } from '@angular/platform-browser';
-import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
 import { ReactiveFormsModule } from '@angular/forms';
-import { AppRoutingModule } from './app-routing.module';
 import { LoginComponent } from './views/login/login.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { ToastrModule } from 'ngx-toastr';
 
 import { MatCardModule } from '@angular/material/card';
-import { MatTabsModule } from '@angular/material/tabs';
 import { MatIconModule } from '@angular/material/icon';
 import { MatListModule } from '@angular/material/list';
 import { MatInputModule } from '@angular/material/input';
 import { MatTableModule } from '@angular/material/table';
 import { MatDialogModule } from '@angular/material/dialog';
+import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 
-import { ListBookComponent } from './views/books/list-book/list-book.component';
-import { BookFormComponent } from './views/books/form-book/book-form.component';
-import { ErrorsInterceptor } from './interceptors/errors-interceptor/errors-interceptor';
-import { AddBookComponent } from './views/books/add-book/add-book.component';
 import { LibraryComponent } from './views/library/library.component';
 import { BookDialogComponent } from './shared/dialog/dialog.component';
+import { AddBookComponent } from './views/library/books/add-book/add-book.component';
+import { ListBookComponent } from './views/library/books/list-book/list-book.component';
+import { BookFormComponent } from './views/library/books/form-book/book-form.component';
+import { ErrorsInterceptor } from './interceptors/errors-interceptor/errors-interceptor';
 
 @NgModule({
   declarations: [
     AppComponent,
     LoginComponent,
-    ListBookComponent,
-    BookFormComponent,
     AddBookComponent,
     LibraryComponent,
+    ListBookComponent,
+    BookFormComponent,
     BookDialogComponent,
   ],
   imports: [
-    CommonModule,
     FormsModule,
-    BrowserModule,
-    HttpClientModule,
-    AppRoutingModule,
-    BrowserAnimationsModule,
-    ToastrModule.forRoot(),
-    ReactiveFormsModule,
+    CommonModule,
     MatListModule,
+    HttpClientModule,
+    ReactiveFormsModule,
+    BrowserAnimationsModule,
 
-    MatInputModule,
-    MatFormFieldModule,
+    ToastrModule.forRoot(),
+
     MatCardModule,
-    MatProgressBarModule,
-    MatTabsModule,
-    MatTableModule,
     MatIconModule,
+    MatInputModule,
+    MatTableModule,
     MatDialogModule,
+    MatButtonModule,
+    MatFormFieldModule,
+    MatProgressBarModule,
   ],
   providers: [
     { provide: HTTP_INTERCEPTORS, useClass: ErrorsInterceptor, multi: true },
