--- src/app/views/books/list-book/list-book.component.html
@@ -1,39 +1,0 @@
-<ng-container>
-  <mat-card>
-    <mat-form-field appearance="standard">
-      <mat-label>Filter</mat-label>
-      <input matInput (keyup)="applyFilter($event)" placeholder="Ex. Robert Martin" #input>
-    </mat-form-field>
-
-    <table mat-table [dataSource]="tableDataSource">
-      <ng-container *ngFor="let column of columns" [matColumnDef]="column.columnDef">
-        <th mat-header-cell *matHeaderCellDef>
-          {{ column.header }}
-        </th>
-        <td mat-cell *matCellDef="let row">
-          {{ column.cell(row) }}
-        </td>
-      </ng-container>
-
-      <ng-container matColumnDef="actions">
-        <mat-header-cell *matHeaderCellDef>Actions</mat-header-cell>
-        <mat-cell *matCellDef="let element">
-          <button mat-icon-button matTooltip="Click to Edit" (click)="edit(element)" class="iconbutton" color="primary">
-            <mat-icon aria-label="Edit">edit</mat-icon>
-          </button>
-          <button mat-icon-button matTooltip="Click to Delete" (click)="remove(element)" class="iconbutton"
-            color="primary">
-            <mat-icon aria-label="Delete">delete_forever</mat-icon>
-          </button>
-        </mat-cell>
-      </ng-container>
-
-      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-      <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
-
-      <tr class="mat-row" *matNoDataRow>
-        <td class="mat-cell" colspan="4">No data matching the filter "{{input.value}}"</td>
-      </tr>
-    </table>
-  </mat-card>
-  <ng-container>
