+++ src/app/views/library/library.component.html
@@ -1,2 +1,1 @@
-<app-add-book></app-add-book>
 <app-list-book></app-list-book>
