+++ src/app/views/login/login.component.html
@@ -7,37 +7,27 @@
             <mat-label>Username</mat-label>
             <input matInput formControlName="username" required />
             <mat-error *ngIf="!formLogin.valid">{{
-              getErrorMessage()
-            }}</mat-error>
+              errorMessage
+              }}</mat-error>
           </mat-form-field>
         </div>
         <div>
           <mat-form-field appearance="fill">
             <mat-label>Password</mat-label>
-            <input
-              matInput
-              formControlName="password"
-              type="password"
-              required
-            />
+            <input matInput formControlName="password" type="password" required />
             <mat-error *ngIf="!formLogin.valid">{{
-              getErrorMessage()
-            }}</mat-error>
+              errorMessage
+              }}</mat-error>
           </mat-form-field>
         </div>
       </div>
-      <input
-        type="reset"
-        class="btn btn-primary mt-4"
-        value="Cancel"
-        onclick=""
-      />
-      <input type="submit" class="btn btn-primary mt-4 right" value="Log In" />
+      <div style="padding-top: 1em">
+        <input type="reset" class="btn btn-primary mt-4" value="Cancel" onclick="" />
+        <input type="submit" class="btn btn-primary mt-4 right" value="Log In" />
+      </div>
     </form>
-    <mat-card-footer *ngIf="isLoading()">
-      <mat-progress-bar
-        style="padding-top: 1em"
-        mode="indeterminate"
-      ></mat-progress-bar> </mat-card-footer
-  ></mat-card>
+    <mat-card-footer *ngIf="isLoading">
+      <mat-progress-bar style="padding-top: 1em" mode="indeterminate"></mat-progress-bar>
+    </mat-card-footer>
+  </mat-card>
 </div>
