+++ src/app/views/login/login.component.ts
@@ -1,6 +1,6 @@
-import { Component, Injectable, OnDestroy, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { timer } from 'rxjs';
+import { Component, Injectable, OnInit } from '@angular/core';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { AccessControl } from 'src/app/shared/accessControl/access-control';
 
 @Injectable({ providedIn: 'root' })
@@ -11,10 +11,10 @@ import { AccessControl } from 'src/app/shared/accessControl/access-control';
 })
 export class LoginComponent implements OnInit {
   formLogin!: FormGroup;
-  loggedStr = 'logged';
-  loading = false;
-
-  readonly time = 1500;
+  
+  private loading = false;
+  private readonly loggedStr = 'logged';
+  private readonly timeBeforeLogIn = 1500;
 
   constructor(private formBuilder: FormBuilder) {}
 
@@ -29,10 +29,10 @@ export class LoginComponent implements OnInit {
     });
   }
 
-  onSubmit(): void {
+  protected onSubmit(): void {
     if (this.formLogin.valid) {
-      this.initLoad();
-      timer(this.time).subscribe(() => {
+      this.startLoad();
+      timer(this.timeBeforeLogIn).subscribe(() => {
         sessionStorage.setItem(this.loggedStr, 'true');
         this.stopLoad();
       });
@@ -43,27 +43,23 @@ export class LoginComponent implements OnInit {
     this.loading = false;
   }
 
-  private initLoad(): void {
+  private startLoad(): void {
     this.loading = true;
   }
 
-  isLoading(): boolean {
+  protected get isLoading(): boolean {
     return this.loading;
   }
 
-  onCancel(): void {
-    this.formLogin.reset(new AccessControl());
-  }
-
-  onLogOut(): void {
+  logOut(): void {
     sessionStorage.removeItem(this.loggedStr);
   }
 
-  isLoggedIn(): boolean {
+  get isLoggedIn(): boolean {
     return sessionStorage.getItem(this.loggedStr) === 'true';
   }
 
-  getErrorMessage(): string {
+  protected get errorMessage(): string {
     return 'You must enter a value';
   }
 }
